auto id implemented
added a for loop that checks each elements id, if there is a gap in the order for example if we deleted a patient then the new id will take its place, if not then the id will be the size of the vector plus 1. this should be scalable and not cause the ids to grow out of hand. this automates the id and the user doesn't needs to worry about it. I will implement a search function, to find the id of a person based on name or date of birth